Remove swap module from main.tf
The terraform apply command got stuck at a point when it was trying to
initialize all the hosts in the cluster. It kept on saying "Waiting for
API server to respond".

Further investigation showed that it was failing on preflight checks:

module.kubernetes.null_resource.kubernetes[0] (remote-exec): [preflight]
Some fatal errors occurred:
module.kubernetes.null_resource.kubernetes[0] (remote-exec):   [ERROR
Swap]: running with swap on is not supported. Please disable swap

This pre-flight check was enabled in kubernetes/kubernetes#47181

Since, the swap was being enabled in main.tf, i have removed that bit to
stop terraform from enabling swap on the host machines.

This closes hobby-kube/guide#25