provisioning: Add etcd_node_count
The current configuration configures an etcd node on each host.

This patch adds etcd_node_count (defaults to 3) and deploys etcd only
on the first N nodes. Disassociating etcd node count from the regular
node count makes it easier to grow the kubernetes without changing the
fault tolerance characteristics of the etcd cluster.

It also  makes it easier to extend the cluster with new nodes with minor
disruption by changing  the node_count variable and performing a terraform
apply.

ref: https://github.com/hobby-kube/guide/issues/49